// seller-app/src/app/seller/models/product/category.model.ts
/* ************************************************************************** */
/*                                                                            */
/*                                                        :::      ::::::::   */
/*   category.model.ts                                  :+:      :+:    :+:   */
/*                                                    +:+ +:+         +:+     */
/*   By: mbah <mbah@student.42lyon.fr>              +#+  +:+       +#+        */
/*                                                +#+#+#+#+#+   +#+           */
/*   Created: 2025/04/24 18:25:12 by mbah              #+#    #+#             */
/*   Updated: 2025/05/05 17:45:04 by mbah             ###   ########.fr       */
/*                                                                            */
/* ************************************************************************** */

export interface Category {
    id: string;
    name: string;
}